selftests/mm: fix charge_reserved_hugetlb.sh test resulting in false errors
Currently, running the selftest we can sometimes observe something like:

	$ ./charge_reserved_hugetlb.sh -cgroup-v2
	...
	write_result is 0
	After write:
	hugetlb_usage=0
	reserved_usage=10485760
	killing write_to_hugetlbfs
	Received 2.
	Deleting the memory
	Detach failure: Invalid argument
	umount: /mnt/huge: target is busy.

Both cases are not actually functionality-wise errors, but rather errors
in the test.

While the unmount error seems to be racy, it will make the test fail:
	$ ./run_vmtests.sh -t hugetlb
	...
	# [FAIL]
	not ok 10 charge_reserved_hugetlb.sh -cgroup-v2 # exit=32

The issue is that we are not waiting for the write_to_hugetlbfs process
to quit. So it might still have a hugetlbfs file open, about which
umount is not happy. Fix that by making "killall" wait for the process
to quit.

The other error ("Detach failure: Invalid argument") does not result in
a test error, but is misleading. Turns out write_to_hugetlbfs.c
unconditionally tries to cleanup using shmdt(), even when. Even worse,
shmaddr is never even set for the SHM case. Fix that as well.

Fixes: 29750f71a9b4 ("hugetlb_cgroup: add hugetlb_cgroup reservation tests")
Cc: Mina Almasry <almasrymina@google.com>
Signed-off-by: David Hildenbrand <david@redhat.com>