virtio-mem: use alloc_contig_range_noprof()
Let's use the noprof variant, this is not an ordinary allocation. We
might free (parts of) these allocated range later using
free_contig_range(), but we might also offline and remove the memory
without ever handing these pages back to the buddy.

Avoid allocation tags for this form of memory offlining.

Signed-off-by: David Hildenbrand <david@redhat.com>