virtio-mem: detected never-onlined vs. previously-onlined correctly across pageblocks
Ever since commit 448b8ec3bf11 ("drivers: virtio_mem: use pageblock size
as the minimum virtio_mem size.") we support (un)plug in pageblock
granularity, not only in (larger) maximum buddy granularity.

We prepared for that in commit 57c5a5b304b0 ("virtio-mem: prepare fake
page onlining code for granularity smaller than MAX_ORDER - 1"), but
missed the case that we can have a range that contains a mixture of
never-onlined and previously-onlined pageblocks within a MAX_PAGE_ORDER
chunk.

This means that we can end up calling free_contig_range() on pages that
were never exposed to the buddy and generic_online_page() on pages that
were already exposed to the buddy.

There are only smaller differences for now -- for example,
generic_online_page() forces the refcounts of all pages to 0, which is
more fragile if the memory was previously exposed to the buddy -- but there
might be bigger differences in the future, for example, once we want to use
compound allocations for alloc_contig_range().

So let's just get it right, and while at it, cleanup and optimize the
pfn+nr_pages handling a bit.

We can trigger one such case by using a virtio-mem device that supports
device blocks <= pageblock size:
 (1) Plug a single device block
 (2) Unplug that device block
 (3) Plug two device blocks
 (4) Unplug two device blocks

During (3) we would not use generic_online_page() on the second device
block that was never exposed to the buddy before.

The expectation is that it should happen rarely and in most cases not cause
real trouble, so let's not CC stable.

Fixes: 57c5a5b304b0 ("virtio-mem: prepare fake page onlining code for granularity smaller than MAX_ORDER - 1")
Signed-off-by: David Hildenbrand <david@redhat.com>