s390/crash_dump: consider device RAM ranges when building elfcore hdr
Currently, virtio-mem device memory will not be included in crashdumps,
because device memory is not part of physmem.

Let's use the new infrastructure to detect these ranges and add them to
the elcore hdr.

We don't really expect to run into the maximum number of memory chunks
right now, but let's catch it so we don't succeed in creating a dump that
partially misses stuff.

Signed-off-by: David Hildenbrand <david@redhat.com>