x86/mm/pat: fix VM_PAT handling during failed fork()
If fork() fails, we might end up having VM_PAT set on a VMA even though
the page tables were not copied yet. Consequently, we might end up
trying to clean up VM_PAT but are unable to obtain the information we
need from the page tables.

Fix it by cleaning up the handling, and simply relying on the src VMA
instead.

Signed-off-by: David Hildenbrand <david@redhat.com>