selftests/mm: gup_tests: option to GUP all pages in a single call
We recently missed detecting a nasty issue during early testing because
the default (!all) tests would not trigger it and even when running
"all" tests it only would happen sometimes.

So let's allow for an easy way to specify "GUP all pages in a single
call", extend the test matrix and extend our default (!all) tests.

Signed-off-by: David Hildenbrand <david@redhat.com>